Dataset: change 'BitWidth' to return an unsigned.
Instead of a uint64_t, which was a pain to cast around and
overkill anyway.  Propagated lots of places.

+++ AbstrConverter.cpp
@@ -166,7 +166,7 @@ size_t AbstrConverter::GetIncoreSize() {
 bool
 AbstrConverter::QuantizeTo8Bit(LargeRAWFile& rawfile,
                                const std::string& strTargetFilename,
-                               uint64_t iComponentSize,
+                               unsigned iComponentSize,
                                uint64_t iSize,
                                bool bSigned,
                                bool bIsFloat,
