+++ IASSConverter.h
@@ -49,7 +49,7 @@ public:
   virtual bool ConvertToRAW(const std::string& strSourceFilename,
                             const std::string& strTempDir,
                             bool bNoUserInteraction,
-                            uint64_t& iHeaderSkip, uint64_t& iComponentSize,
+                            uint64_t& iHeaderSkip, unsigned& iComponentSize,
                             uint64_t& iComponentCount, bool& bConvertEndianess,
                             bool& bSigned, bool& bIsFloat,
                             UINT64VECTOR3& vVolumeSize,
@@ -60,7 +60,7 @@ public:
   /// @todo unimplemented!
   virtual bool ConvertToNative(const std::string& strRawFilename,
                                const std::string& strTargetFilename,
-                               uint64_t iHeaderSkip, uint64_t iComponentSize,
+                               uint64_t iHeaderSkip, unsigned iComponentSize,
                                uint64_t iComponentCount, bool bSigned,
                                bool bFloatingPoint,
                                UINT64VECTOR3 vVolumeSize,
