+++ NRRDConverter.cpp
@@ -54,7 +54,7 @@ NRRDConverter::NRRDConverter()
 
 bool NRRDConverter::ConvertToRAW(const std::string& strSourceFilename,
                                  const std::string& strTempDir, bool,
-                                 uint64_t& iHeaderSkip, uint64_t& iComponentSize,
+                                 uint64_t& iHeaderSkip, unsigned& iComponentSize,
                                  uint64_t& iComponentCount,
                                  bool& bConvertEndianess, bool& bSigned,
                                  bool& bIsFloat, UINT64VECTOR3& vVolumeSize,
@@ -444,7 +444,7 @@ bool NRRDConverter::ConvertToRAW(const std::string& strSourceFilename,
 }
 
 bool NRRDConverter::ConvertToNative(const std::string& strRawFilename, const std::string& strTargetFilename, uint64_t iHeaderSkip,
-                             uint64_t iComponentSize, uint64_t iComponentCount, bool bSigned, bool bFloatingPoint,
+                             unsigned iComponentSize, uint64_t iComponentCount, bool bSigned, bool bFloatingPoint,
                              UINT64VECTOR3 vVolumeSize,FLOATVECTOR3 vVolumeAspect, bool bNoUserInteraction,
                              const bool bQuantizeTo8Bit) {
