+++ VFFConverter.cpp
@@ -49,7 +49,7 @@ VFFConverter::VFFConverter()
 
 bool VFFConverter::ConvertToRAW(
   const std::string& strSourceFilename, const std::string&, bool,
-  uint64_t& iHeaderSkip, uint64_t& iComponentSize, uint64_t& iComponentCount,
+  uint64_t& iHeaderSkip, unsigned& iComponentSize, uint64_t& iComponentCount,
   bool& bConvertEndianess, bool& bSigned, bool& bIsFloat,
   UINT64VECTOR3& vVolumeSize, FLOATVECTOR3& vVolumeAspect,
   std::string& strTitle, std::string& strIntermediateFile,
@@ -182,7 +182,7 @@ bool VFFConverter::ConvertToRAW(
 }
 
 bool VFFConverter::ConvertToNative(const std::string& strRawFilename, const std::string& strTargetFilename, uint64_t iHeaderSkip,
-                             uint64_t iComponentSize, uint64_t iComponentCount, bool bSigned, bool bFloatingPoint,
+                             unsigned iComponentSize, uint64_t iComponentCount, bool bSigned, bool bFloatingPoint,
                              UINT64VECTOR3 vVolumeSize,FLOATVECTOR3 vVolumeAspect, bool, bool bQuantizeTo8Bit) {
 
   // create header textfile from metadata
