+++ VGStudioConverter.cpp
@@ -51,7 +51,7 @@ VGStudioConverter::VGStudioConverter()
 
 bool VGStudioConverter::ConvertToRAW(const std::string& strSourceFilename,
                             const std::string&, bool,
-                            uint64_t& iHeaderSkip, uint64_t& iComponentSize,
+                            uint64_t& iHeaderSkip, unsigned& iComponentSize,
                             uint64_t& iComponentCount, bool& bConvertEndianess,
                             bool& bSigned, bool& bIsFloat,
                             UINT64VECTOR3& vVolumeSize,
